+++ lib/couchrest/model/design.rb
@@ -120,8 +120,6 @@ module CouchRest
       def load_from_database(db = database, id = nil)
         id ||= self['_id']
         db.get(id)
-      rescue CouchRest::NotFound
-        nil
       end
 
       # Calculate and update the checksum of the Design document.